requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
Django==1.11.7
django-extensions==1.9.7
djangorestframework==3.7.3
gcloud==0.18.3
google==1.9.3
google-api-python-client==1.6.4
googleapis-common-protos==1.5.3
googlemaps==2.5.1
httplib2==0.10.3
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
oauth2client==4.1.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
